+++ Shogi1.Domain/Model/AIs/IAI.cs
@@ -10,6 +10,6 @@ namespace Shogi1.Domain.Model.AIs
         /// </summary>
         /// <param name="board">盤</param>
         /// <returns>(指し手、評価値)</returns>
-        (MoveBase moveBase, double eval) DecideMove(Board board);
+        (MoveBase moveBase, int eval) DecideMove(Board board);
     }
 }
